+++ R/religion.R
@@ -15,7 +15,7 @@
 #'   Buddhist  \tab   489,030,000 \tab 7.08 \% \cr
 #'   Folk      \tab   406,140,000 \tab 5.88 \% \cr
 #'   Other     \tab    59,330,000 \tab 0.86 \% \cr
-#'   Jewish    \tab    15,670,000 \tab 0.23 \% 
+#'   Jewish    \tab    15,670,000 \tab 0.23 \% \cr
 #' }
 #'
 #' @inheritParams r_sample_factor
@@ -34,3 +34,28 @@ religion <- hijack(r_sample_factor,
     x = c("Christian", "Muslim", "None", "Hindu", "Buddhist", "Folk", "Other", "Jewish"),
     prob = c(0.31477, 0.23163, 0.16323, 0.14985, 0.07083, 0.05882, 0.00859, 0.00227)
 )
+
+## pacman::p_load(dplyr, curl, qdapRegex)
+##
+## m <- read.csv(curl("https://raw.githubusercontent.com/ramnathv/data/gh-pages/all/pew_religion_religionByCountry.csv"))
+## 
+## m[, -1] <- lapply(m[, -1], function(x) {
+##     as_numeric(sapply(rm_number(x, extract= TRUE), paste, collapse=""))
+## })
+## 
+## 
+## c(apply(m[, -1], 2, sum)) %>%
+##     vect2df("Religion", "Total") %>%
+##     mutate(
+##         Proportion = Total/sum(Total),
+##         Percent = paste(round(100*Proportion, 2), "\\%"),
+##         Proportion = round(Proportion, 5),
+##         Total = format(Total,  big.mark = ",", scientific=FALSE),
+##         Religion = gsub("^U.+", "None", gsub("\\..+", "", Religion))
+##     ) %>%
+##     arrange(desc(Proportion)) %>%
+##     #select(1, 3) %>% dput
+##     select(-3) %>%
+##     tabular() %>%
+##     cat(file="clipboard")
+