+++ tests/testthat/test-browser.R
@@ -1,0 +1,7 @@
+context("Checking browser")
+
+test_that("browser ...",{
+
+
+})
+